add createNewNode method and (some) tests
With the 'Node' class implemented as an internal helper class, there was
no way to get an instance of a 'node'.  This was particulary problematic
for unit testing purposes.  The 'Node' class has several methods that
should be included in a unit test.

To enable access to a 'node' instance, we added a 'createNewNode' method
of the linked-list class that returns a new Node object.  This is
useful for not only enabling unit testing, but also makes the
instantiation of new nodes within the linked-list (insert) methods a bit
more descriptive.